[CHERI_CSA] ProvenanceSourceChecker: silence for hybrid mode
