+++ clang/lib/StaticAnalyzer/Checkers/CHERI/CHERIUtils.h
@@ -1,0 +1,26 @@
+//===-- CHERIUtils.h  -------------------------------------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_LIB_STATICANALYZER_CHECKERS_CHERI_CHERIUTILS_H
+#define LLVM_CLANG_LIB_STATICANALYZER_CHECKERS_CHERI_CHERIUTILS_H
+
+#include "clang/StaticAnalyzer/Core/Checker.h"
+
+namespace clang {
+namespace ento {
+namespace cheri {
+
+bool isPureCapMode(const ASTContext &C);
+
+bool isPointerToCapTy(const QualType Type, ASTContext &Ctx);
+
+} // namespace cheri
+} // namespace ento
+} // namespace clang
+
+#endif // LLVM_CLANG_LIB_STATICANALYZER_CHECKERS_CHERI_CHERIUTILS_H
